wip 3/0/8/0, pending pragma lazystring, pragma private, constexpr folds, other misc syntax/grammar updates, all t7. does not match release build (yet)

+++ T7CompilerLib/OpCodes/T7OP_GetLocalFunction.cs
@@ -42,7 +42,7 @@ namespace T7CompilerLib.OpCodes
             uint JumpTo = After.CommitAddress + After.GetSize();
             uint JumpFrom = CommitAddress + GetSize();
 
-            ((short)(JumpTo - JumpFrom)).GetBytes(Endianess).CopyTo(data, GetCommitDataAddress());
+            ((int)(JumpTo - JumpFrom)).GetBytes(Endianess).CopyTo(data, GetCommitDataAddress());
         }
 
         public override uint GetCommitDataAddress()
